+++ DataSets/DataSets/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Data;
 
 namespace DataSets
 {
@@ -10,40 +9,24 @@ namespace DataSets
             using (var db = new DB("Data Source=.\\;Initial Catalog=DataSetsDB;Integrated Security=True"))
             {
                 Console.WriteLine("SELECT *:");
-                printEmailValues(db);
+                db.printEmailValues();
                 Console.WriteLine();
 
                 Console.WriteLine("After INSERT:");
                 db.InsertIntoEmailTable();
-                printEmailValues(db);
+                db.printEmailValues();
                 Console.WriteLine();
 
                 Console.WriteLine("After UPDATE:");
                 db.UpdateEmailTable();
-                printEmailValues(db);
+                db.printEmailValues();
                 Console.WriteLine();
 
                 Console.WriteLine("After DELETE:");
                 db.DeleteFromEmailTable();
-                printEmailValues(db);
+                db.printEmailValues();
                 Console.WriteLine();
             }
-        }
-
-        private static void printEmailValues(DB db)
-        {
-            var sqlReader = db.ReadEmailTable();
-
-            while (sqlReader.Read())
-            {
-                readSingleRow(sqlReader);
-            }
-
-            sqlReader.Close();
-        }
-        private static void readSingleRow(IDataRecord record)
-        {
-            Console.WriteLine($"email id: {record[0]}, email: {record[1]}, lecturer id: {record[2]}");
         }
     }
 }
